loader: Move verity-hash.h to include
