+++ grub-core/loader/i386/efi/linux.c
@@ -28,7 +28,7 @@
 #include <grub/efi/efi.h>
 #include <grub/tpm.h>
 
-#include "../verity-hash.h"
+#include <grub/verity-hash.h>
 
 GRUB_MOD_LICENSE ("GPLv3+");
