+++ grub-core/loader/i386/linux.c
@@ -37,7 +37,8 @@
 #include <grub/linux.h>
 #include <grub/tpm.h>
 
-#include "verity-hash.h"
+#include <grub/verity-hash.h>
+
 GRUB_MOD_LICENSE ("GPLv3+");
 
 #ifdef GRUB_MACHINE_PCBIOS
