+++ include/grub/verity-hash.h
